Workspaces are their own module, not sandbox plumbing
Move the Workspace base and RepoWorkspace into a new druks.workspaces
module. Both sat in the wrong layer: the base lived in
druks.sandbox.datastructures although it resolves OAuth, folds the MCP
registry, and commits the DB — execution policy, not VM plumbing — while
RepoWorkspace was buried inside the ship app, out of reach of any other app.

Now they share one home. Apps subclass and import from druks.workspaces, and
sandbox.datastructures keeps only its data shapes. Behavior is unchanged: both
classes move verbatim, and workflows, agents, and the apps switch to the new
import path. druks.workspaces stays out of the pinned author surface — first-
party apps borrow it directly, the same way contrib borrows other platform
machinery.
